+++ src/bot/main.cpp
@@ -6,7 +6,7 @@
 /*   By: ysabik <ysabik@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/26 01:09:47 by ysabik            #+#    #+#             */
-/*   Updated: 2024/05/28 02:32:16 by ysabik           ###   ########.fr       */
+/*   Updated: 2024/05/28 08:32:10 by ysabik           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -55,6 +55,7 @@ int main(int argc, char **argv) {
 		.addExecutor(new PrivBotCommand("PRIVMSG", List<std::string>("M")("MSG")))
 		.addExecutor(new PingExecutor("!ping", List<std::string>("!p")))
 		.addExecutor(new PongExecutor("!pong", List<std::string>("!pp")))
+		.addExecutor(new JokeExecutor("!joke", List<std::string>("!j")))
 		;
 
 	std::string	usage = "Usage: " + std::string(argv[0]) + " <IPv4> <port> [<password>]";